LPD-91213 Write scan insights in the SEO Studio service
