--- workspaces/liferay-seostudio-workspace/client-extensions/liferay-seostudio-etc-crawler/src/main/java/com/liferay/seo/studio/model/DetectorResult.java
@@ -1,29 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.seo.studio.model;
-
-/**
- * @author Brooke Dalton
- */
-public class DetectorResult {
-
-	public DetectorResult(String errorMessage, String state) {
-		_errorMessage = errorMessage;
-		_state = state;
-	}
-
-	public String getErrorMessage() {
-		return _errorMessage;
-	}
-
-	public String getState() {
-		return _state;
-	}
-
-	private final String _errorMessage;
-	private final String _state;
-
-}
