--- workspaces/liferay-seostudio-workspace/client-extensions/liferay-seostudio-etc-crawler/src/main/java/com/liferay/seo/studio/model/Domain.java
@@ -1,25 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.seo.studio.model;
-
-import org.json.JSONObject;
-
-/**
- * @author Brooke Dalton
- */
-public class Domain {
-
-	public Domain(JSONObject jsonObject) {
-		_hostname = jsonObject.optString("hostname", null);
-	}
-
-	public String getHostname() {
-		return _hostname;
-	}
-
-	private final String _hostname;
-
-}
