+++ workspaces/liferay-seostudio-workspace/client-extensions/liferay-seostudio-etc-crawler/src/main/java/com/liferay/seo/studio/service/SEOStudioService.java
@@ -7,17 +7,27 @@ package com.liferay.seo.studio.service;
 
 import com.liferay.client.extension.util.spring.boot3.client.LiferayOAuth2AccessTokenManager;
 import com.liferay.client.extension.util.spring.boot3.service.BaseService;
+import com.liferay.petra.function.transform.TransformUtil;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.util.ListUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.seo.studio.model.CrawlHit;
-import com.liferay.seo.studio.model.Domain;
 
 import java.net.URI;
 
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
+
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
+
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 
 import org.json.JSONArray;
 import org.json.JSONObject;
@@ -32,6 +42,42 @@ import org.springframework.web.util.UriComponentsBuilder;
  */
 @Component
 public class SEOStudioService extends BaseService {
+
+	public void addSEOStudioScanInsights(
+			long accountEntryId, JSONObject insightJSONObject,
+			long seoStudioScanId)
+		throws Exception {
+
+		JSONArray pageURLsJSONArray = insightJSONObject.optJSONArray(
+			"pageURLs");
+
+		if ((pageURLsJSONArray == null) || pageURLsJSONArray.isEmpty()) {
+			return;
+		}
+
+		List<String> pageURLs = TransformUtil.transform(
+			pageURLsJSONArray.toList(), object -> (String)object);
+
+		_addSEOStudioScanInsights(
+			accountEntryId, insightJSONObject, pageURLs,
+			_resolveSEOStudioPageIdsMap(
+				accountEntryId, pageURLs, seoStudioScanId),
+			seoStudioScanId);
+	}
+
+	public JSONObject fetchSEOStudioDomainJSONObject(long seoStudioDomainId) {
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/domains/" + seoStudioDomainId
+		).build();
+
+		String responseJSON = get(_getAuthorization(), uriComponents.toUri());
+
+		if (Validator.isNull(responseJSON)) {
+			return null;
+		}
+
+		return new JSONObject(responseJSON);
+	}
 
 	public String getActiveSEOStudioScans() {
 		UriComponents uriComponents = UriComponentsBuilder.fromPath(
@@ -79,38 +125,6 @@ public class SEOStudioService extends BaseService {
 		return crawlHits;
 	}
 
-	public Domain getSEOStudioDomain(long seoStudioDomainId) {
-		UriComponents uriComponents = UriComponentsBuilder.fromPath(
-			"/o/seo-studio/domains/" + seoStudioDomainId
-		).build();
-
-		String responseJSON = get(_getAuthorization(), uriComponents.toUri());
-
-		if (Validator.isNull(responseJSON)) {
-			return null;
-		}
-
-		return new Domain(new JSONObject(responseJSON));
-	}
-
-	public String getSEOStudioPages(
-		int page, int pageSize, long seoStudioScanId) {
-
-		UriComponents uriComponents = UriComponentsBuilder.fromPath(
-			"/o/seo-studio/pages"
-		).queryParam(
-			"filter",
-			"r_seoStudioScanToSEOStudioPages_seoStudioScanId eq '" +
-				seoStudioScanId + "'"
-		).queryParam(
-			"page", page
-		).queryParam(
-			"pageSize", pageSize
-		).build();
-
-		return get(_getAuthorization(), uriComponents.toUri());
-	}
-
 	public String patchSEOStudioDomain(
 		JSONObject jsonObject, long seoStudioDomainId) {
 
@@ -120,6 +134,17 @@ public class SEOStudioService extends BaseService {
 
 		return patch(
 			_getAuthorization(), jsonObject.toString(), uriComponents.toUri());
+	}
+
+	public long getSEOStudioDomainId(JSONObject seoStudioScanJSONObject) {
+		long seoStudioScanRunId = seoStudioScanJSONObject.getLong(
+			"r_seoStudioScanRunToSEOStudioScans_seoStudioScanRunId");
+
+		JSONObject seoStudioScanRunJSONObject = new JSONObject(
+			_getSEOStudioScanRun(seoStudioScanRunId));
+
+		return seoStudioScanRunJSONObject.getLong(
+			"r_seoStudioDomainToSEOStudioScanRuns_seoStudioDomainId");
 	}
 
 	public String patchSEOStudioScan(
@@ -145,33 +170,6 @@ public class SEOStudioService extends BaseService {
 		jsonObject.put("state", state);
 
 		return patchSEOStudioScan(jsonObject, seoStudioScanId);
-	}
-
-	public String postSEOStudioInsightType(JSONObject jsonObject) {
-		UriComponents uriComponents = UriComponentsBuilder.fromPath(
-			"/o/seo-studio/insight-types"
-		).build();
-
-		return post(
-			_getAuthorization(), jsonObject.toString(), uriComponents.toUri());
-	}
-
-	public String postSEOStudioPagesBatch(JSONArray jsonArray) {
-		UriComponents uriComponents = UriComponentsBuilder.fromPath(
-			"/o/seo-studio/pages/batch"
-		).build();
-
-		return post(
-			_getAuthorization(), jsonArray.toString(), uriComponents.toUri());
-	}
-
-	public String postSEOStudioScanInsightsBatch(JSONArray jsonArray) {
-		UriComponents uriComponents = UriComponentsBuilder.fromPath(
-			"/o/seo-studio/scan-insights/batch"
-		).build();
-
-		return post(
-			_getAuthorization(), jsonArray.toString(), uriComponents.toUri());
 	}
 
 	public URI toCrawlURI(String hostname) {
@@ -210,6 +208,85 @@ public class SEOStudioService extends BaseService {
 		return "seo_studio_" + seoStudioDomainId;
 	}
 
+	private void _addSEOStudioPages(
+		long accountEntryId, List<String> pageURLs, long seoStudioScanId) {
+
+		for (int i = 0; i < pageURLs.size(); i += _BATCH_SIZE) {
+			JSONArray seoStudioPagesJSONArray = new JSONArray();
+
+			List<String> batchPageURLs = ListUtil.subList(
+				pageURLs, i, i + _BATCH_SIZE);
+
+			for (String pageURL : batchPageURLs) {
+				seoStudioPagesJSONArray.put(
+					_toSEOStudioPageJSONObject(
+						accountEntryId, pageURL, seoStudioScanId));
+			}
+
+			_postSEOStudioPagesBatch(seoStudioPagesJSONArray);
+		}
+	}
+
+	private void _addSEOStudioScanInsights(
+		long accountEntryId, JSONObject insightJSONObject,
+		List<String> pageURLs, Map<String, Long> seoStudioPageIdsMap,
+		long seoStudioScanId) {
+
+		JSONObject seoStudioInsightTypeJSONObject = new JSONObject(
+			_postSEOStudioInsightType(
+				_toSEOStudioInsightTypeJSONObject(
+					accountEntryId, insightJSONObject, seoStudioScanId)));
+
+		long seoStudioInsightTypeId = seoStudioInsightTypeJSONObject.getLong(
+			"id");
+
+		String detectedDateString = Instant.now(
+		).truncatedTo(
+			ChronoUnit.SECONDS
+		).toString();
+
+		for (int i = 0; i < pageURLs.size(); i += _BATCH_SIZE) {
+			JSONArray seoStudioScanInsightsJSONArray = new JSONArray();
+
+			List<String> batchPageURLs = ListUtil.subList(
+				pageURLs, i, i + _BATCH_SIZE);
+
+			for (String pageURL : batchPageURLs) {
+				Long seoStudioPageId = seoStudioPageIdsMap.get(pageURL);
+
+				if (seoStudioPageId == null) {
+					if (_log.isWarnEnabled()) {
+						_log.warn("Unable to get a page for URL " + pageURL);
+					}
+
+					continue;
+				}
+
+				seoStudioScanInsightsJSONArray.put(
+					_toSEOStudioScanInsightJSONObject(
+						accountEntryId,
+						insightJSONObject.getString("classification"),
+						detectedDateString, seoStudioInsightTypeId,
+						seoStudioPageId, seoStudioScanId));
+			}
+
+			if (seoStudioScanInsightsJSONArray.isEmpty()) {
+				continue;
+			}
+
+			_postSEOStudioScanInsightsBatch(seoStudioScanInsightsJSONArray);
+		}
+
+		if (_log.isInfoEnabled()) {
+			_log.info(
+				StringBundler.concat(
+					"Posted ", pageURLs.size(), " ",
+					insightJSONObject.getString("name"),
+					" SEO Studio scan insights for SEO Studio insight type ID ",
+					seoStudioInsightTypeId));
+		}
+	}
+
 	private String _getAuthorization() {
 		return _liferayOAuth2AccessTokenManager.getAuthorization(
 			"liferay-seostudio-etc-crawler-oahs");
@@ -236,6 +313,200 @@ public class SEOStudioService extends BaseService {
 
 		return get(_getAuthorization(), uriComponents.toUri());
 	}
+
+	private Map<String, Long> _getSEOStudioPageIdsMap(long seoStudioScanId) {
+		Map<String, Long> seoStudioPageIdsMap = new HashMap<>();
+
+		int page = 1;
+
+		while (true) {
+			JSONArray itemsJSONArray = new JSONObject(
+				_getSEOStudioPages(page, 2000, seoStudioScanId)
+			).optJSONArray(
+				"items"
+			);
+
+			if ((itemsJSONArray == null) || itemsJSONArray.isEmpty()) {
+				break;
+			}
+
+			for (Object object : itemsJSONArray) {
+				JSONObject seoStudioPageJSONObject = (JSONObject)object;
+
+				seoStudioPageIdsMap.put(
+					seoStudioPageJSONObject.getString("pageURL"),
+					seoStudioPageJSONObject.getLong("id"));
+			}
+
+			page++;
+		}
+
+		return seoStudioPageIdsMap;
+	}
+
+	private String _getSEOStudioPages(
+		int page, int pageSize, long seoStudioScanId) {
+
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/pages"
+		).queryParam(
+			"filter",
+			"r_seoStudioScanToSEOStudioPages_seoStudioScanId eq '" +
+				seoStudioScanId + "'"
+		).queryParam(
+			"page", page
+		).queryParam(
+			"pageSize", pageSize
+		).build();
+
+		return get(_getAuthorization(), uriComponents.toUri());
+	}
+
+	private String _getSEOStudioScanRun(long seoStudioScanRunId) {
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/scan-runs/" + seoStudioScanRunId
+		).build();
+
+		return get(_getAuthorization(), uriComponents.toUri());
+	}
+
+	private String _postSEOStudioInsightType(JSONObject jsonObject) {
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/insight-types"
+		).build();
+
+		return post(
+			_getAuthorization(), jsonObject.toString(), uriComponents.toUri());
+	}
+
+	private String _postSEOStudioPagesBatch(JSONArray jsonArray) {
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/pages/batch"
+		).build();
+
+		return post(
+			_getAuthorization(), jsonArray.toString(), uriComponents.toUri());
+	}
+
+	private String _postSEOStudioScanInsightsBatch(JSONArray jsonArray) {
+		UriComponents uriComponents = UriComponentsBuilder.fromPath(
+			"/o/seo-studio/scan-insights/batch"
+		).build();
+
+		return post(
+			_getAuthorization(), jsonArray.toString(), uriComponents.toUri());
+	}
+
+	private Map<String, Long> _resolveSEOStudioPageIdsMap(
+			long accountEntryId, List<String> pageURLs, long seoStudioScanId)
+		throws Exception {
+
+		Map<String, Long> seoStudioPageIdsMap = _getSEOStudioPageIdsMap(
+			seoStudioScanId);
+
+		List<String> missingPageURLs = ListUtil.filter(
+			pageURLs, pageURL -> !seoStudioPageIdsMap.containsKey(pageURL));
+
+		if (ListUtil.isEmpty(missingPageURLs)) {
+			return seoStudioPageIdsMap;
+		}
+
+		_addSEOStudioPages(accountEntryId, missingPageURLs, seoStudioScanId);
+
+		long deadline = System.currentTimeMillis() + 60000;
+
+		while (true) {
+			Set<String> existingPageURLs = seoStudioPageIdsMap.keySet();
+
+			if (existingPageURLs.containsAll(pageURLs)) {
+				break;
+			}
+
+			if (System.currentTimeMillis() > deadline) {
+				if (_log.isWarnEnabled()) {
+					_log.warn(
+						"Timed out waiting for pages to be readable for SEO " +
+							"Studio scan ID " + seoStudioScanId);
+				}
+
+				break;
+			}
+
+			Thread.sleep(1000);
+
+			seoStudioPageIdsMap.putAll(
+				_getSEOStudioPageIdsMap(seoStudioScanId));
+		}
+
+		return seoStudioPageIdsMap;
+	}
+
+	private JSONObject _toSEOStudioInsightTypeJSONObject(
+		long accountEntryId, JSONObject insightJSONObject,
+		long seoStudioScanId) {
+
+		return new JSONObject(
+		).put(
+			"category", insightJSONObject.getString("category")
+		).put(
+			"description", insightJSONObject.getString("description")
+		).put(
+			"externalReferenceCode",
+			insightJSONObject.getString("name") + "_" + seoStudioScanId
+		).put(
+			"fixHint", insightJSONObject.getString("fixHint")
+		).put(
+			"name", insightJSONObject.getString("name")
+		).put(
+			"r_accountToSEOStudioInsightTypes_accountEntryId", accountEntryId
+		).put(
+			"r_seoStudioScanToSEOStudioInsightTypes_seoStudioScanId",
+			seoStudioScanId
+		).put(
+			"severity", insightJSONObject.getString("severity")
+		);
+	}
+
+	private JSONObject _toSEOStudioPageJSONObject(
+		long accountEntryId, String pageURL, long seoStudioScanId) {
+
+		return new JSONObject(
+		).put(
+			"pageURL", pageURL
+		).put(
+			"r_accountToSEOStudioPages_accountEntryId", accountEntryId
+		).put(
+			"r_seoStudioScanToSEOStudioPages_seoStudioScanId", seoStudioScanId
+		);
+	}
+
+	private JSONObject _toSEOStudioScanInsightJSONObject(
+		long accountEntryId, String classification, String detectedDateString,
+		long seoStudioInsightTypeId, long seoStudioPageId,
+		long seoStudioScanId) {
+
+		return new JSONObject(
+		).put(
+			"classification", classification
+		).put(
+			"detectedDate", detectedDateString
+		).put(
+			"r_accountToSEOStudioScanInsights_accountEntryId", accountEntryId
+		).put(
+			"r_seoStudioInsightTypeToScanInsights_seoStudioInsightTypeId",
+			seoStudioInsightTypeId
+		).put(
+			"r_seoStudioPageToSEOStudioScanInsights_seoStudioPageId",
+			seoStudioPageId
+		).put(
+			"r_seoStudioScanToSEOStudioScanInsights_seoStudioScanId",
+			seoStudioScanId
+		);
+	}
+
+	private static final int _BATCH_SIZE = 100;
+
+	private static final Log _log = LogFactory.getLog(SEOStudioService.class);
 
 	@Autowired
 	private LiferayOAuth2AccessTokenManager _liferayOAuth2AccessTokenManager;
